freebsd: cross-compile workspace cleanly on x86_64-unknown-freebsd
- errno: align freebsd_errno API with linux (GetErrno trait, MAX:u16,
  init(i64)); type S::* constants/predicates as u32 to match the
  cross-platform Mode width; inline libuv UV_E* fallback constants
  (windows_sys is windows-only)
- sys: route errno_ptr via __error(); cast Mode→libc::mode_t at every
  chmod/mkdir/umask boundary (FreeBSD mode_t is u16); add
  posix::sysctlbyname/timeval, c::{sysctlbyname,getloadavg,sockaddr_dl,
  UTIME_NOW}, time::timestamp, eventfd(2); declare lchmod extern;
  hardcode DT_WHT=14
- stat: libc flattens st_*timespec into st_*time/st_*time_nsec on the
  BSDs (only Apple keeps the nested struct)
- aio: build kevent via zeroed-then-assign helper so the optional ext[4]
  field (FreeBSD ≥12 ABI) doesn't break struct-literal init; udata is
  *mut c_void
- threading/Futex: port _umtx_op WAIT_UINT_PRIVATE/WAKE_PRIVATE directly
  against libc::{_umtx_time,_umtx_op}
- io: enable adt_const_params for the kqueue ApplyAction const-generic;
  fix EINTR/kevent tag spelling
- runtime: getdirentries(2) for the FreeBSD dir iterator (libc binds
  neither getdents nor getdirentries); declare getcontext extern; widen
  st_mode at S::IS*/fchmod call sites; sa_family_t is u8 on BSD
- cares_sys: EAI_NODATA was dropped by RFC 3493 — use historical 7
- bun_core: FreeBSD backtrace() is size_t/size_t; pthread_getthreadid_np
  for debug thread_id; add strings::contains_case_insensitive_ascii

cargo check --workspace --target x86_64-unknown-freebsd: 0 errors
cargo check --workspace (host linux): 0 errors

+++ src/io/openForWriting.rs
@@ -140,7 +140,7 @@ where
                     *pollable = true;
                 }
 
-                *is_socket = (stat.st_mode as u32 & bun_sys::S::IFMT) == bun_sys::S::IFSOCK;
+                *is_socket = (stat.st_mode as u32 & bun_sys::S::IFMT as u32) == bun_sys::S::IFSOCK as u32;
 
                 if force_sync || isatty {
                     // Prevents interleaved or dropped stdout/stderr output for terminals.
